requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.8
distlib==0.3.8
dj-database-url==2.2.0
dj-static==0.0.6
Django==5.0.6
django-celery-results==2.5.1
django-filter==24.2
django-log-request-id==2.1.0
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.14.0
gunicorn==22.0.0
hiredis==2.3.2
identify==2.5.36
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
nodeenv==1.8.0
numpy==2.0.1
packaging==24.1
phone_gen==2.4.10
phonenumbers==8.13.38
phonenumberslite==8.13.38
platformdirs==4.2.2
pre-commit==3.7.1
prompt_toolkit==3.0.47
psycopg2==2.9.9
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
random-username==1.0.2
redis==5.0.6
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
scipy==1.14.0
six==1.16.0
sqlparse==0.5.0
static3==0.7.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
whitenoise==6.7.0
